Release/v0.6.0

+++ module/category-tree/src/Infrastructure/Handler/DeleteTreeCommandHandler.php
@@ -12,6 +12,8 @@ namespace Ergonode\CategoryTree\Infrastructure\Handler;
 use Ergonode\CategoryTree\Domain\Command\DeleteTreeCommand;
 use Ergonode\CategoryTree\Domain\Entity\CategoryTree;
 use Ergonode\CategoryTree\Domain\Repository\TreeRepositoryInterface;
+use Ergonode\Core\Infrastructure\Exception\ExistingRelationshipsException;
+use Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface;
 use Webmozart\Assert\Assert;
 
 /**
@@ -24,11 +26,20 @@ class DeleteTreeCommandHandler
     private $repository;
 
     /**
-     * @param TreeRepositoryInterface $repository
+     * @var RelationshipsResolverInterface
      */
-    public function __construct(TreeRepositoryInterface $repository)
-    {
+    private $relationshipsResolver;
+
+    /**
+     * @param TreeRepositoryInterface        $repository
+     * @param RelationshipsResolverInterface $relationshipsResolver
+     */
+    public function __construct(
+        TreeRepositoryInterface $repository,
+        RelationshipsResolverInterface $relationshipsResolver
+    ) {
         $this->repository = $repository;
+        $this->relationshipsResolver = $relationshipsResolver;
     }
 
     /**
@@ -36,10 +47,22 @@ class DeleteTreeCommandHandler
      *
      * @throws \Exception
      */
-    public function __invoke(DeleteTreeCommand $command)
+    public function __invoke(DeleteTreeCommand $command): void
     {
         $categoryTree = $this->repository->load($command->getId());
-        Assert::isInstanceOf($categoryTree, CategoryTree::class, sprintf('Can\'t find category tree with id "%s"', $command->getId()));
+        Assert::isInstanceOf(
+            $categoryTree,
+            CategoryTree::class,
+            sprintf(
+                'Can\'t find category tree with ID "%s"',
+                $command->getId()
+            )
+        );
+
+        $relationships = $this->relationshipsResolver->resolve($command->getId());
+        if (!$relationships->isEmpty()) {
+            throw new ExistingRelationshipsException($command->getId());
+        }
 
         $this->repository->delete($categoryTree);
     }
