+++ module/workflow/src/Infrastructure/Handler/Workflow/DeleteWorkflowCommandHandler.php
@@ -9,7 +9,10 @@ declare(strict_types = 1);
 
 namespace Ergonode\Workflow\Infrastructure\Handler\Workflow;
 
+use Ergonode\Core\Infrastructure\Exception\ExistingRelationshipsException;
+use Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface;
 use Ergonode\Workflow\Domain\Command\Workflow\DeleteWorkflowCommand;
+use Ergonode\Workflow\Domain\Entity\Workflow;
 use Ergonode\Workflow\Domain\Repository\WorkflowRepositoryInterface;
 use Webmozart\Assert\Assert;
 
@@ -23,20 +26,40 @@ class DeleteWorkflowCommandHandler
     private $repository;
 
     /**
-     * @param WorkflowRepositoryInterface $repository
+     * @var RelationshipsResolverInterface
      */
-    public function __construct(WorkflowRepositoryInterface $repository)
-    {
+    private $relationshipsResolver;
+
+    /**
+     * @param WorkflowRepositoryInterface    $repository
+     * @param RelationshipsResolverInterface $relationshipsResolver
+     */
+    public function __construct(
+        WorkflowRepositoryInterface $repository,
+        RelationshipsResolverInterface $relationshipsResolver
+    ) {
         $this->repository = $repository;
+        $this->relationshipsResolver = $relationshipsResolver;
     }
 
     /**
      * @param DeleteWorkflowCommand $command
+     *
+     * @throws ExistingRelationshipsException
      */
     public function __invoke(DeleteWorkflowCommand $command)
     {
         $workflow = $this->repository->load($command->getId());
-        Assert::notNull($workflow);
+        Assert::isInstanceOf(
+            $workflow,
+            Workflow::class,
+            sprintf('Can\'t find workflow with ID "%s"', $command->getId())
+        );
+
+        $relationships = $this->relationshipsResolver->resolve($command->getId());
+        if (!$relationships->isEmpty()) {
+            throw new ExistingRelationshipsException($command->getId());
+        }
 
         $this->repository->delete($workflow);
     }
